+++ application/models/Tracker/Sites/DokiFansubs.php
@@ -17,7 +17,7 @@ class DokiFansubs extends Base_Site_Model {
 		];
 	}
 
-	public function getTitleData(string $title_url, bool $firstGet = FALSE) {
+	public function getTitleData(string $title_url, bool $firstGet = FALSE) : ?array {
 		$fullURL = $this->getFullTitleURL($title_url);
 		return $this->parseFoolSlide($fullURL, $title_url);
 	}