+++ application/models/Tracker/Sites/DynastyScans.php
@@ -51,7 +51,7 @@ class DynastyScans extends Base_Site_Model {
 		return $chapterData;
 	}
 
-	public function getTitleData(string $title_url, bool $firstGet = FALSE) {
+	public function getTitleData(string $title_url, bool $firstGet = FALSE) : ?array {
 		$titleData = [];
 
 		$fullURL = $this->getFullTitleURL($title_url);