+++ application/models/Tracker/Sites/MangaCow.php
@@ -15,7 +15,7 @@ class MangaCow extends Base_Site_Model {
 		];
 	}
 
-	public function getTitleData(string $title_url, bool $firstGet = FALSE) {
+	public function getTitleData(string $title_url, bool $firstGet = FALSE) : ?array {
 		$titleData = [];
 
 		$fullURL = $this->getFullTitleURL($title_url);