+++ application/models/Tracker/Sites/MangaPanda.php
@@ -16,7 +16,7 @@ class MangaPanda extends Base_Site_Model {
 		];
 	}
 
-	public function getTitleData(string $title_url, bool $firstGet = FALSE) {
+	public function getTitleData(string $title_url, bool $firstGet = FALSE) : ?array {
 		$titleData = [];
 
 		$fullURL = $this->getFullTitleURL($title_url);